combine calculator classes into one
